Move check_satpy to satpy.utils
