+++ satpy/tests/reader_tests/test_eps_l1b.py
@@ -28,7 +28,7 @@ import pytest
 import satpy
 import xarray as xr
 from satpy.readers import eps_l1b as eps
-from satpy._config import config_search_paths
+from satpy._config import get_config_path
 from satpy.tests.utils import make_dataid
 
 grh_dtype = np.dtype([("record_class", "|i1"),
@@ -43,7 +43,7 @@ grh_dtype = np.dtype([("record_class", "|i1"),
 def create_sections(structure):
     """Create file sections."""
     sections = {}
-    format_fn = list(config_search_paths("eps_avhrrl1b_6.5.xml"))[-1]
+    format_fn = get_config_path("eps_avhrrl1b_6.5.xml")
     form = eps.XMLFormat(format_fn)
     for count, (rec_class, sub_class) in structure:
         try:
