+++ satpy/tests/test_config.py
@@ -22,28 +22,6 @@ import unittest
 from unittest import mock
 
 
-class TestCheckSatpy(unittest.TestCase):
-    """Test the 'check_satpy' function."""
-
-    def test_basic_check_satpy(self):
-        """Test 'check_satpy' basic functionality."""
-        from satpy._config import check_satpy
-        check_satpy()
-
-    def test_specific_check_satpy(self):
-        """Test 'check_satpy' with specific features provided."""
-        from satpy._config import check_satpy
-        with mock.patch('satpy._config.print') as print_mock:
-            check_satpy(readers=['viirs_sdr'], extras=('cartopy', '__fake'))
-            checked_fake = False
-            for call in print_mock.mock_calls:
-                if len(call[1]) > 0 and '__fake' in call[1][0]:
-                    self.assertNotIn('ok', call[1][1])
-                    checked_fake = True
-            self.assertTrue(checked_fake, "Did not find __fake module "
-                                          "mentioned in checks")
-
-
 class TestBuiltinAreas(unittest.TestCase):
     """Test that the builtin areas are all valid."""
